Intutionize fvex in iset.mm
Unfortunately (given how much fvex is used in set.mm) it seems
a bit unlikely that there's a way to replace the case
elimination with something else. Showing that the function
value exists in the usual case (where F is a function and
we are within its domain) is easy, though.

So this involves renaming fvex to funfvex and adding a
( Fun F /\ A e. dom F ) condition.

Also add funfveu which is a lemma for this proof.