Remove tz6.12i and consequences from iset.mm
It seems unlikely we'd be able to make this "value of a class
is either in the range or is the empty set" stuff work when
we don't have case elimination. Fortunately, it is relatively
lightly used in set.mm.

The theorems removed here are: tz6.12i , fvbr0 , fvrn0 , and
fvssunirn .