Intutionize dffn5 in iset.mm
Change biconditional to implication and rename to dffn5im .

My attempts to prove the converse all led to ( F ` x ) e. _V
and I'm not sure whether F = ( x e. A |-> ( F ` x ) ) constrains
F in such a way as to ensure that the domain of F is A (as
opposed to a subset thereof). We, of course, do not have the
ability to do case elimination based on whether ( F ` x ) is
empty or non-empty which is basically what set.mm does here.