Intuitionize fvun1 in iset.mm
This proof combines the fvun and fvun1 proofs from set.mm. fvun
doesn't work as written because you would need to know whether
you are dealing with an element of dom F or of dom G (and fvun1
answers that: dom F ).

For this reason, remove fvun . With the exception of undifixp
(which we can worry about later), it is only used in set.mm
via fvun1 . So there is no need to worry about whether we
want a variant which combines fvun1 and fvun2 by adding a
condition like X e. A \/ X e. B .