Remove suppss and suppssr from iset.mm
In both cases the current proof depends on contraposition
( necon1ad or similar), and this might not be fixable.
Although these theorems are used in set.mm , the first
use is for Cantor Normal Form, and at least at a quick
glance, we have a bunch of other issues to deal with
before we figure out what (if anything) to do with
Cantor Normal Form constructively.