Intuitionize rexrnmpt in iset.mm
Instead of translating E. to A. via dfrex2 , prove in a way
analogous to rexrnmpt . Towards the end we seem to be lacking
the analogue to ralimiaa which we would need, and so there
ends up being a number of logic steps involving pm5.32 , exbi ,
and the like. But it does work.