Intuitionize fprg in iset.mm
Instead of proving it from fpr and the weak deduction theorem
(which we don't have), prove it with a proof similar to the
fpr proof but relying on rnsnopg instead of rnsnop (and likewise
for all the other theorems we use). Interestingly, the resulting
proof is 407 characters (from SHOW PROOF/SIZE), compared with
740 characters for the set.mm proof (which relies on fpr and
applies the weak deduction theorem).