Intuitionize csbov and friends in iset.mm
This is a combination of set.mm before Aug 2018 and latest set.mm.

Remove csbov (unused in set.mm) and csbov123 (replaced by csbov123g ).
Rename csbovgOLD to csbov123g .
Change D to V in csbov12g (just a class variable rename, the
meaning of the theorem is the same).