Remove some caov* theorems iset.mm
These are theorems which rely on ovex (which is not something we
have and which doesn't seem constructive in the sense
that it handles both the cases in which the operands are in
the domain and the case where they are not). The difference
between these (for example, caov4 ) and the ones which work
(for example, caov4d ) is mainly that the ones which work
tend to have a closure hypothesis.

The theorems removed are caov4 , caov411 , caov42 , caovdir ,
caovdilem , and caovlem2 .