Add relelfvdm to iset.mm
This is like elfvdm in set.mm but also requires that F be a
relation (which is likely to be true most of the time we
would need this, I would imagine).

The proof follows from elfv and releldm and what seems like a
surprisingly large amount of substitution and other logic
(but which isn't doing anything especially tricky).