Remove offn and consequences from iset.mm
The intuitionizing needed (assuming we really need to add
conditions that the operations are functions being evaluated
within their domains) is extensive. Although these theorems
are used a moderate amount in set.mm, much of it is later
in the file.

The theorems being removed are offn , offveq , caofid0l ,
caofid0r , caofid1 , and caofid2 .