Intuitionize ot3rdg in iset.mm
In set.mm, ot3rdg only has a condition that C exist. We need
one for A and B as well (at least for the straightforward proof).
So add that condition and rename to ot3rdgg (not sure there is
an especially good way to name this one given the "change the
name when it diverges from set.mm" rule).