Add resfunexgALT to ofmresex to iset.mm
In set.mm these are in a section called "Finite induction (for
finite ordinals)" (although I'm not sure they depend on
finite induction). Some other parts of that section are in
iset.mm already, but these ones seem to have slipped through
the cracks.

This is a copy from set.mm without changes; follow up changes
will be needed.