Intuitionize brtpos in iset.mm
Rename to brtposg and add the conditions that A and B exist.
The proof is taken from the A e. _V /\ B e. _V case of the
set.mm brtpos proof, with the additional changes that
opex is changed to opexg and opswap is changed to opswapg .

Add opswap to mmil.raw.html list of set.mm theorems missing
in iset.mm because it wasn't there.