Expand the intuitionization document in mmil.html
One more example of a removing distinct variable constraints with
a temporary variable.

Point at theorems which show what disjunctive syllogism is and
what we have for it.

Say that we aren't guaranteeing that everything on the "set.mm
theorems we don't have" is unprovable (just that most are
unprovable).

Add df-ex , mo2 , dfrex2 , moabsn , and snex to that list.