Remove zero/successor/limit theorems from iset.mm
This is tz7.44-1 , tz7.44-2 , and tz7.44-3 . In the words of
[Crosilla], we cannot make "case distinctions such as in the
classical ordinal cases of successor and limit" so the whole
idea of these theorems doesn't work (nor do the mechanics,
being based on "if" which we don't have in iset.mm).

Also remove tz7.44lem1 as it wouldn't be of much use unless
we were doing case elimination on empty/successor/limit.