Remove non-intuitionistic recursion from iset.mm
The first part of transfinite recursion (tfr2a and its lemmas
basically) works as-is. The other part does not. Although it
should be intuitionizable in some form, the current proofs
depend on ordinal trichotomy and other theorems that are
impossible (or likely impossible), as well as more mundane
issues like fvex and opex .

The theorems removed are tfr1a , tfr2b , tfr1 , tfr2 , tfr3 ,
recsfnon , and recsval .

The lemmas removed are tfrlem9a , tfrlem10 , tfrlem11 , tfrlem12 ,
tfrlem13 , tfrlem14 , tfrlem15 , tfrlem16 .

Add text to mmil.raw.html which explains this situation and also
the constrast between df-recs and df-rdg .