Add the first thema of the 4 Stoic logic themata
Stoic logic at its basics had five axiomatic arguments
(the indemonstratables) and four inference rules (the themata).

This commit adds proofs of the first thema (of the four themata).
We already have the indemonstrables.

The idea is to show that the axiomatic system used in MPE is
at least as strong as this ancient logic system.
We can prove this if we can prove (or accept) its axioms and rules.
The statements and results are interesting
in their own right.  They may also be useful in other theorems,
since these were "obvious" enough to be considered the basis of a
logic system thousands of years ago.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>