Improved mmj2 workaround for setvar
mmj2 doesn't currently know about "setvar"; it expects the older
"set" constant in some checks inside "macros/definitionCheck.js".
We've worked around that by modifying set.mm.  However,
it's not really a good practice to modify the file you're checking
before you check it.

This commit modifies the source code of mmj2's
"macros/definitionCheck.js" after downloading it so it will look for
"setvar" instead of "set".  That means we can check set.mm
without modifying it.

NOTE: This will currently cause a failure because iset.mm
uses "set" instead of "setvar".  Normally iset.mm tries to
be consistent with set.mm, so the solution is to modify iset.mm
to use the same convention.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>