Documentation fixes to iset.mm
Change the guidelines on naming to allow theorems with the same
name but more hypotheses in iset.mm (such as snex or opex). Add
paragraph about generally trying to keep set.mm and iset.mm in
sync.

Rewrap comments.

Revise the comments in tfri1 , tfri2 , and tfri3 concerning the
condition that the recursion rule has to be defined "everywhere".