Document how to convert $d to F/_
Mario Carneiro posted on the Metamath mailing list
on Tue, 28 May 2019 16:34:40 -0400, a general approach
for converting $d statements into F/_ statements
(see "Re: [Metamath] Replacing $d with F/ and F/_").

This commit adds that information to the "conventions"
discussion about freeness, so that others will be able to
find that information later.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>