Add fnoa to iset.mm
The proof ends up being longer than the set.mm one because we
have additional steps to show that we are evaluating functions
within their domains. But because the characteristic function
( z e. _V |-> suc z ) is defined on all sets, it is relatively
straightforward (by contrast, fnom may require a version of
rdgifnon which only requires a characteristic function defined
on ordinals, or some other approach).