Add onsucelsucr to iset.mm
This might be the closest we can get to ordsucelsuc . At
least at first glance it would appear that trying to prove
A e. B -> suc A e. suc B runs into needing "either suc A equals B
or suc A is an element of B". Maybe induction or something
else could prove that direction.

Add ordsucelsuc and ordsucsssuc to the list of missing theorems
on the web page.