Document ordinals in iset.mm
Add sucon to iset.mm (this is easily proved now).

Describe why we don't have We, Fr, and if (at least
not currently).

Describe (on mmil.html) ordinal theorems from set.mm which are not
in iset.mm. In many cases the notes are taken directly from
Mario Carneiro's comments on github. The editing is mostly
to account for theorems which have since been proved,
statements which have been retracted, etc.