Remove reuiotaex from iset.mm
It is unused and apparently when I added it I was unaware of
riotacl which is similar but stronger.

Change the missing theorems list to point to riotacl for set.mm's
riotaex .

Add a variety of theorems to the missing theorems list which
happened to be nearby in the git logs.