Fix finite recursion in iset.mm
Most seriously, the definition df-frec had been saying that
g e. _om but in fact g is a function from natural numbers
to sets. I guess we could say g e. ~P ( _om X. _V ) but
that wouldn't help prove that classes exist, so just go
for g e. _V .

Along with that, it is convenient for df-frec to say that
m e. _om . This might not be completely necessary, but
it makes the proofs easier and matches the intended
values for m .

Fix frec0g and frecfnom for the changed df-frec definition.

Add frectfr lemma to make it easier to apply tfri1d or tfri2d when
proving finite recursion theorems.

Add frecsuc including lemmas frecabex , frecsuclem1 , frecsuclemdm ,
frecsuclem2 , and frecsuclem3 .

Update mmil.html accordingly.