Add om0x to mmil.html
Certainly the set.mm proof relies on case elimination and given
the nature of the theorem it sort of seems likely that other
proofs would too. Even in set.mm, perhaps it should be marked
discouraged (for reasons described in its comment), but I don't
know if that is controversial given some of the things set.mm
has done with other "simplify via case elimination" situations.