Intuitionize ecopover and friends in iset.mm
The theorems ecopovsym , ecopovtrn , and ecopover have hypotheses
which, basically, rely on evaluating functions outside their
domains. Introduce new versions ecopovsymg , ecopovtrng , and
ecopoverg (perhaps not brilliantly named, but I didn't think
of anything better) which have hypothoses with conditions
on the variables that they are elements of ` S ` .

The original theorems still hold, so I guess we can keep them
for set.mm parity, but it is sort of a hypothetical parity in
that I'm pretty sure we'll not be using them (based on looking
at set.mm uses).