Document that we prefer to minimize axioms
In practice we prefer to minimize axioms, but nowhere do we seem
to document that.  Modify the "conventions" text to expressly
document this preference, and why we do it.

I modified the existing axiom of choice text to do this, because
the axiom of choice is an especially obvious case of this.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>