Modify df-rq in iset.mm
This is based on df-rq from the 2012 set.mm with the y e. Q.
condition added. My guess is that this is equivalent to the
one without the y e. Q. condition, or the df-rq from
2019 set.mm, but I wasn't able to quickly prove that, so at
least for now adopt a definition which is (at least for
current uses) easier to work with.