Add npsspw and elinp to iset.mm
npsspw is taken from part of the npex proof and is used there and
for elinp.

As for elinp it is similar to set.mm's elnp but for our different
definition of Dedekind cut. I'm assuming it will be most useful to
have a version of this with two class variables for the lower and
upper sets, rather than using 1st and 2nd a lot.