Fix .travis.yml, regen-discouraged, add iset-discouraged, fixes #1120
Tweak .travis.yml per issue #1120 from @benjub;
the file had gotten a little cruft and no-longer-needed statements.

This adds iset-discouraged, so iset.mm undergoes the same
"discourage checking" that set.mm already undergoes.
If you *do* want to use something that is discouraged in iset.mm,
you can use this to regenerate the file:

> scripts/regen-discouraged --results iset-discouraged iset.mm

This commit also fixes scripts/regen-discouraged; it was
supposed to already work this way, but due to a bug it didn't
actually write to the --results file.  That's now fixed.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>