Remove discouraged tags from iset.mm real numbers
This is for the section on real number construction. The current
thinking is that it will be better not to have the tags while we
work on the construction. This is to avoid work in terms of
regenerating iset-discouraged and also so that the minimizer
can make use of these theorems.

Once the construction is complete and we have introduced NN ,
QQ , etc, and the various theorems which use them, we might have
reason to add the discouraged tags back.