Add sotricim , sotritric and nqtric to iset.mm
We already have trichotomy for positive fractions, so this just
adds one more form of it (in the style of set.mm's sotric ).

We can think of sotricim and sotritric as lemmas although they
are also useful in general to supply our nearest approximation to
set.mm's sotric .

Update mmil.html accordingly.