Fix date of Fenton's trf1ALT to 22-April-2011
Theorem tfr1ALT was claimed to be contributed by Scott Fenton
on 17-Aug-1994, but that is clearly an error
(it wasn't in the repo at that time).

Change its contribution date to 22-Apr-2011, the best available
estimate by Norman Megill per a private email from him
(Subject "Re: Unusually early set.mm contribution: tfr1ALT"
sent Tue, 1 Oct 2019 10:44:30 -0400).

This means that the earliest contribution recorded in set.mm
by someone other than Norman Megill is 1999-09-05,
a proof shortening by David Harvey of "notnot2".
This appears to be accurate information.

We want all credits and dates to be accurate, of course.
That sais, fixing this one is especially important for the
Gource visualization of set.mm's history, because that
visualization highlights early contributions by others.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>