Add basic infastructure for "people/" subdirectory
Add the basics for a new "people/" subdirectory, which
will store public avatars of contributors (where the contributor
has indicated that such a public avatar is okay).

The immediate use is for "Gource" visualizations, but there
may other occasions where an avatar may be useful.

Note: Avatars are an image of the contributor's choosing;
they do not need to be a face (and often aren't).
We do not require contributors to give their real names, to respect
people's privacy; contributors may use a pseudonym of their choosing.
Of course, we may choose to not accept a pseudonym or avatar
for our purposes :-).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>