Use 'finds' for Metamath 100 #74, fixes #1159
Straighten out peculiarities in the entry for
Metamath 100 proof # 74 (chronologically 2/100), by
using 'finds' (by Norman Megill, 1995-04-14) as the primary proof,
but also noting theorem 'findes' (by Raph Levien, 9-Jul-2003) and
'nnind' (by Norman Megill, 10-Jan-1997).

We had been listing 'findes', but theorem 'finds' came first,
so it makes more sense to list 'finds' instead.

If this commit is accepted, it implies that:
* the Google docs spreadsheet for Metamath 100 will need modification,
  for entry 2 it becomes Norman Megill, 1995-04-14, 'finds'.
* Freek will need to be told that in his list entry #74 needs
  'Raph Levien' changed to 'Norman Megill'.
If this commit is accepted I plan to do that.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>