set.mm: Add markers for Metamath 100 proofs < 50
This adds markers within set.mm for Metamath 100 proofs
with ids less than 50.  This will help prevent renames
in set.mm that aren't reflected in the Metamath 100 list,
and may be of interest to readers of the respective theorems.

This partly implements this issue:
https://github.com/metamath/set.mm/issues/1162

I have intentionally *not* rewrapped, because I want to
avoid changing anything else and possibly triggering a
painful merge conflict resolution. The plan is rewrap later
after all these markers are merged.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>