Fix lag in gource visualization, and tweak some captions
This fixes a lag problem noticed by Benoit.
The gource animations take some time and that *partly* explains the lag,
but when I rechecked the settings I realized they produced a *big* lag.

I wanted the visualization to be a short video, not a movie, so I use:
> seconds-per-day=0.01
That's fine by itself, but I originally had this as a setting:
> max-file-lag=1

This meant that a new file's display in the visualization could be delayed
by up to 100 days, and that was the problem. I set them to match:
> max-file-lag=0.01
and regenerated the video. The "appearance" video takes time,
but it starts on time & I think it more accurately visualizes the data.

If you look at the video you'll see that every second
is actually 50 days, not 100.
I generate the videos at 60 fps (the default), then encode them at 30 fps.
That way, I see the videos in half the time. I do this because every
time I tweak a configuration I need to regenerate it & see it;
double-speed generation makes my process faster.
I've seen a *lot* of versions of this video :-).
I documented this effect in comments of scripts/gourcify.

While we're at it, I fixed some capitalizations in the captions.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>