mmset.raw.html: Add getting started link
Nowhere on the set.mm front page do we provide a hint of
how to start contributing something else.  That should be fixed.

This adds a single link on the top right to the GitHub page on
"Getting started with contributing (to Metamath set.mm)".
That will at least help people find that information.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>