mmdeduction.raw.html: Change tilde to logical not
In the "email section" the tilde is used to represent logical not.
Replace tilde with the actual Unicode logical not character, which
is clearer.

In the long term I think this email section should be rewritten, e.g.,
use ` ... ` expressions, use ph and ps instead of A and B, and s on.
That would make this material consistent with
everything else.  But this tweak at least means that we no longer have
a long list of warnings and problems.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>