mmil.raw.html: "prnadd: prnaddl" should be "prnmadd: prnmaddl"
The hypertext link "prnaddl" doesn't work because there is
no such label in iset.mm.  As a result, it fails a check.
The page also refers to the set.mm label "prnadd" that doesn't exist.

Think that the *intended* entry is that set.mm's "prnmadd" somewhat
maps to iset.mm's "prnmaddl".  This commit fixes the table.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>