.travis.yml: Also verify set.mm using mmverify.py
The mmverify.py verifier is clear but extremely slow.
We currently use it for iset.mm, but iset.mm is much smaller.

This commit adds another job to verify set.mm using mmverify.py.
To prevent it from taking too long, we'll verify set.mm but stop it
when it gets to the label "mathbox".  That way we'll at least
verify the body of set.mm.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>