Reword df-if comment
Remove text about weak deduction theorem. Add mention of
decidability.

Although the example "let f(x)=0 if x=0 and 1/x otherwise." isn't
a great one constructively (as the real number version of that
would be based on apartness), I don't have a better one immediately,
so leave it in for now.