mmset.raw.html: Add link to Gource visualization
The Gource visualization is a good way to show that
there have been many contributors to set.mm over time.
However, there's been no link to it in mmset.raw.html.

This commit adds a link to it from the "Theorem Sampler" section,
as that seems like a plausible place to add the link.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>