mmset.raw.html: Remove Ghilbert link
Ghilbert is not very active, and in any case it doesn't
make sense to place it at the top of mmset.raw.html.
So, remove its link from mmset.raw.html.

Note: this does *not* remove all references to Ghilbert.
Indeed, we don't *want* to remove all references.
Ghilbert continues to be mentioned under "Related Links"
on mmrecent.html, and is also noted in the Metamath book.
The idea is that those other references should suffice,
we just don't need the reference here as well.

This makes the "external links" a single item for now, but
I think that's fine.  No doubt we'll add other external links
in the future, so I chose to leave the heading in plural form.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>