Modify mm_100.html now that Cayley-Hamilton's proved
Modify mm_100.html now that the Cayley-Hamilton Theorem
has been proved by Alexander van der Vekens on 2019-11-25.
This Metamath 100 proof #49, and is the 72nd Metamath 100 proof.

While we're at it, this modifies mm_100.html to add some
additional information about Metamath for those who might be
coming to this page for the first time.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>