mm_100.html: Add "fourier" (total of 73)
Modify mm_100.html to record the successful formalization of the
Fourier Series proof by Glauco Siliprandi.
It is proven as "fourier" on 2019-12-11. This is Metamath 100 #76;
it is the 73rd completed proof.

While we're at it, I made a few more tweaks. For example, I added a
hyperlink to the Gource visualization.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>