Update 19.16 , 19.17 , 19.19 , 19.21-2 in iset.mm
Because these are unused in iset.mm we don't need to worry about
keeping around the old versions (if anyone has branches/forks of
iset.mm and is inconvenienced by this, feel free to suggest a
more gradual transition).

- Copy 19.16 , 19.17 , and 19.19 from set.mm (and leave them at
the same place in the file where they were).
- Add comment to 19.21h (it was already discouraged, but say what
to use instead).
- Copy 19.21-2 from set.mm and move it later in the file (so it
is after 19.21 which the proof uses).