Modify signed real definition in set.mm
Change df-plr to define the addition directly rather than being
in terms of +pR .

Add prsrlem1 and addsrmo which are lemmas which basically
replace th3q and its lemmas.

Rewrite addsrpr in terms of addsrmo not th3q .

Regenerate dmaddsr proof (the proof doesn't change in terms of
which theorems are applied in which order, but the changed
df-plr definition means the proof is not identical).