Remove ovec and related theorems from set.mm
Because of the large number of hypotheses and (related) fact that
it is doing a lot, ovec was a bit unwieldy.

Also, the use of a separate function for the pre-operation can
be replaced by theorems such as addsrmo and mulsrmo . Getting rid
of pre-operations means that th3q , th3qlem1 , th3qlem2 , and
th3qcor are no longer needed.

Previous changes made these theorems unused, so this change
just removes them.