Add mulcmpblnr to iset.mm
Includes lemma mulcmpblnrlemg (taken from mulcmpblnrlem from set.mm
but with an additional antecedent).  The proof of mulcmpblnrlemg
closely follows the set.mm proof but adds a whole bunch of
"something e. P." conditions to a lot of steps.

The mulcmpblnr proof basically follows the steps in the set.mm
proof but has to deal with the additional antecedent in
mulcmpblnrlemg .