Shorten set.mm proof of mulcmpblnr
The logic in the mulcmpblnr proof was arranged in an unusual way,
including one propositional logic theorem - rnlem - which is
used nowhere else in set.mm.  At first I thought maybe rnlem
and the rest of the logic here was worth it to keep the proof
from being too long, but rewriting the proof using unimaginative
propositional logic (syl2anc and friends, chiefly) actually makes
it shorter.

It goes from 874 characters in SHOW PROOF/SIZE to 859 characters.