Reword complex number axiom comments in iset.mm
Change "axiom xxx of 22 for" to "axiom for" - the numbering changed
because (a) the trichotomy axiom was removed and replaced by two
axioms, and (b) we haven't yet added axiom(s) to replace the
supremum axiom.

Change "ZF set theory" to "set theory". The particular flavor of
set theory (CZF or IZF or whatever) isn't especially relevant
in this context.