Add 2pwuninelg to iset.mm
This corresponds to 2pwuninel in set.mm - the additional A e. V
condition is fine for what we want this for. This proof could
be used to shorten the proof of 2pwuninel in set.mm but would
use an additional axiom, so I won't propose it for set.mm.

Remove ovec from mmil.html as it no longer exists in set.mm.