Add mul02 to iset.mm
The proof is based on the proof of mul01i from set.mm from 2012.
The statement of the theorem is the same as current set.mm.

Add mul02lem2 with a proof in terms of mul02 .  Adjust the comment
to say that it is corollary, not a lemma, but keep it in because
it has a number of usages in set.mm.

Add mul01 , mul02i , mul01i, mul02d and mul01d .  These are as
in set.mm except for some renaming of hypotheses to reflect
their different position in the file.

Remove all of these from mmil.html because they are now proved.