Improve linebreak documentation in conventions
We already had some text in conventions about line length.
Improve it based on the recent experience in sqr->sqrt.

There's a risk of "saying so much the reader can't remember anything".
The amount of text here is intentially brief, so I think
that risk isn't too bad, and if information is not documented then
people are very likely to not find it :-).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>