Add addge0 through ltneg to iset.mm
All theorems are stated as in set.mm.  The proofs are copied
unmodified except ltsub1 and ltsub2 which get new proofs similar
to lesub1 and lesub2 respectively.

Add this pattern to mmil.html - it is harder in iset.mm than
set.mm to prove a theorem for <_ and get the one for < for free
(analogous to A. and E. ).