Add inelr and rimul to iset.mm
The inelr proof is adapted from set.mm but changes not equal
to zero to apart from zero and also uses tightness of apartness.

The statement of rimul is as in set.mm . The proof is generally
similar, but (1) not-equal is changed to apart, and (2) division
is replaced with reciprocal (since we plan on proving all the
division and reciprocal theorems later, but we do have a bunch of
multiplication ones).