Add condition that reciprocal is positive
This is in the real number axioms. The proof that we had for the
existence of the reciprocal already constructed a number which
was clearly positive, so we just need to expand that proof
slightly to state that it is positive.

Since axprecex was lightly used, just change all the former
usages of it.