Improve scripts to help rename labels
It can be tricky to rename Metamath labels.
In concept it's easy, but there are tricky issues
(e.g., if a label includes "." and you use sed to replace them,
the "." may match "any character").
Naively creating naming changes can miss things
(e.g., you may catch a $p without its matching $e's even
when the $e's have the same label as a prefix).

This commit updates the scripts scripts/mass-rename and
scripts/rename-proposals to improve their automation of
renaming labels. Since nothing can be perfect, we continue to
break the problem into two different steps:

* scripts/rename-proposals : Creates renaming proposals.
  You can then hand-tweak these proposals unti you're happy.
* scripts/mass-rename : Implements renaming (usually you'd
  apply this to the hand-tweaked proposals).

This commit adds a number of options (including --help)
and improves the automation to try to maximally do
"the right thing".

Of course, any result of renaming would need to go through the
usual review process, but we want to maximize the likelihood
of it being right in the first place.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>