Replace remulap0 with mulap0r in iset.mm
Generalize from real numbers to complex numbers. Basically the
same proof works, we just need to change remulext to mulext .

Improve the naming for better consistency with theorems like
mulne0 in iset.mm.

Change the two usages of remulap0 to use mulap0r instead. At
least for now, that means they need extra steps to say a real
number is a complex number. Generally speaking that seems to
be the approach in set.mm for complex number theorems - it is
better to provide recnd and whatever other tools to show that
real numbers are complex, than to provide every theorem with
both real and complex versions.

In that spirit, get rid of remulext (which is now unused) in
favor of mulext . We still have remulext1 and remulext2 , but
those are part of the proof of mulext so they are a different case.